requirements.txt
msal==1.31.0
requests==2.32.3
markdown==3.7
python-dotenv==1.0.1
